Move tests under RequestTests
